Add a guide for dumping the game using a USB storage device (#1471)
* Update README dumping requirements and instructions
Include the option of using a USB storage device for dumping game files in the README.

* Add note to hard drive dumping instructions
Clarify that the instructions for dumping game files directly from the Xbox 360 hard drive differ from those for using a USB storage device by adding a note to the hard drive dumping guide.

* Add dumping guide for USB storage devices
Add a comprehensive guide for dumping the game using a USB storage device instead of a Hard Drive Transfer Cable.

* Apply suggestions from the review

Co-authored-by: Hyper <34012267+hyperbx@users.noreply.github.com>

* Implement additional suggestions from the review
- Rename the file to ```DUMPING-USB-en.md```.
- Add Xbox 360 E to the note about incompatible consoles.
- Move the note under the Pre-requisites section.

---------

Co-authored-by: Hyper <34012267+hyperbx@users.noreply.github.com>